requirements.txt
requests
hachoir
xxhash
zstd
Pillow
python-i18n

